Move the ODataStringEscapeOption to Json subfolder
