+++ src/Microsoft.OData.Core/Build.Net35/Microsoft.OData.Core.NetFX35.csproj
@@ -379,6 +379,9 @@
     <Compile Include="$(ODataCrossTargettingSourcePath)\Json\ODataJsonWriterUtils.cs">
       <Link>Microsoft\OData\Core\Json\ODataJsonWriterUtils.cs</Link>
     </Compile>
+    <Compile Include="$(ODataCrossTargettingSourcePath)\Json\ODataStringEscapeOption.cs">
+      <Link>Microsoft\OData\Core\Json\ODataStringEscapeOption.cs</Link>
+    </Compile>
     <Compile Include="$(ODataCrossTargettingSourcePath)\Json\TextWriterWrapper.cs">
       <Link>Microsoft\OData\Core\Json\TextWriterWrapper.cs</Link>
     </Compile>
@@ -813,9 +816,6 @@
     </Compile>
     <Compile Include="$(ODataCrossTargettingSourcePath)\ODataStreamReferenceValue.cs">
       <Link>Microsoft\OData\Core\ODataStreamReferenceValue.cs</Link>
-    </Compile>
-    <Compile Include="$(ODataCrossTargettingSourcePath)\ODataStringEscapeOption.cs">
-      <Link>Microsoft\OData\Core\ODataStringEscapeOption.cs</Link>
     </Compile>
     <Compile Include="$(ODataCrossTargettingSourcePath)\ODataTypeAnnotation.cs">
       <Link>Microsoft\OData\Core\ODataTypeAnnotation.cs</Link>
